# backend2/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BBWS Sumatera VIII API Reference",
    "version": "1.0.0",
    "description": "Dokumentasi REST API BBWS Sumatera VIII Hono Engine (Auto-generated)"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Server Lokal"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/balai/bbwssumatera8/api/kesiapsiagaan-bencana": {
      "get": {
        "tags": [
          "Kesiapsiagaan-bencana"
        ],
        "summary": "GET /balai/bbwssumatera8/api/kesiapsiagaan-bencana",
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        }
      }
    },
    "/balai/bbwssumatera8/api/kesiapsiagaan-bencana/{id}": {
      "get": {
        "tags": [
          "Kesiapsiagaan-bencana"
        ],
        "summary": "GET /balai/bbwssumatera8/api/kesiapsiagaan-bencana/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        }
      },
      "put": {
        "tags": [
          "Kesiapsiagaan-bencana"
        ],
        "summary": "PUT /balai/bbwssumatera8/api/kesiapsiagaan-bencana/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Kesiapsiagaan-bencana"
        ],
        "summary": "DELETE /balai/bbwssumatera8/api/kesiapsiagaan-bencana/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/balai/bbwssumatera8/api/kesiapsiagaan-bencana/upload": {
      "post": {
        "tags": [
          "Kesiapsiagaan-bencana"
        ],
        "summary": "POST /balai/bbwssumatera8/api/kesiapsiagaan-bencana/upload",
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  }
}